+++ packages/plugin-types/src/deep_partial.ts
@@ -1,7 +1,11 @@
 /**
- * Makes all the properties of the entire nested object partial
- * instead of just the top level properties
+ * Makes all the properties of the entire nested object partial for Amplify generated configs
+ * instead of just the top level properties. Other properties are not changed.
  */
-export type DeepPartial<T> = {
-  [P in keyof T]?: DeepPartial<T[P]>;
+export type DeepPartialAmplifyGeneratedConfigs<T> = {
+  [P in keyof T]?: P extends 'auth' | 'data' | 'storage'
+    ? T[P] extends object
+      ? DeepPartialAmplifyGeneratedConfigs<T[P]>
+      : Partial<T[P]>
+    : T[P];
 };